Move CompetitionsImageDataLoaderPresentationAdapter to separate file
